Better address abort checkout and minor fixes
Main problem with express checkout is the redirect made by the server
where the context is ajax.

The previous solution works for Plus gateway because the order is aborted
at order review page, after the payment, but for the ecs isn't possible
because the payment is done at checkout.

So I refactor a bit the CheckoutDropper class.

Additionally I fixed some issues, probably caused by the alignment of the
develop branch.

- WooCommerce instance isn't injected into ExpressCheckout gateway nor into
  SharedRepository, lead to errors when we try to get the user agent.
- Order key not sanitized during Ipn call
- Wrong sanitization for the amount to `int` instead of consider it as a
  string. Consider it a `float` or `int` could lead to problems because of
  the decimal and thousand separators.
- OrderId returned by `wc_get_order_id_by_order_key` isn't an integer
  as they claim, but a string. So I cast it.
- Wrong Order update logic after payment execution, after converted
  the old `if/else` logic to `switch` I missed to set the correct conditional
  for pui leading to an error where the order isn't update at all.

PPP-328

+++ src/Order/OrderFactory.php
@@ -35,7 +35,8 @@ class OrderFactory
         assert(is_string($orderKey));
 
         $orderKey = $this->orderKeyByJson($orderKey);
-        $orderId = wc_get_order_id_by_order_key($orderKey);
+        // Cast to int because the function return a int as string.
+        $orderId = (int)wc_get_order_id_by_order_key($orderKey);
         $order = $this->createById($orderId);
 
         $this->bailIfInvalidOrder($order);
@@ -46,7 +47,7 @@ class OrderFactory
     /**
      * Create and order by the given Id
      *
-     * @param $orderId
+     * @param int $orderId
      * @return WC_Order|WC_Order_Refund
      * @throws RuntimeException
      */
